Detect rewinds when consumer has any rewind in history
Rewinds can occur even when the consumer has no lag and its often an indication
that something that gone quite wrong. This change checks for a rewind when there
is any current lag. If a rewind is found in the history of the consumer offsets,
it then checks to see if the consumer has 'recovered' from the rewind. If not,
it is marked as a rewind. If it has recovered, we fall back to the usual
usual status evaluation rules.

Currently the wiki doesn't say anything about REWIND status, but does say:
>  If any lag within the window is zero, the status is considered to be OK.

However that misses the case where lag can be zero for a while and then jump
back in time. This can occur during a rebalance as consumers drain existing
work (for those that keep a pipeline of work) or due to a broker bug that throws an
Exception when attempting to retrieve the current offsets, causing many consumers
to revert to the beginning of the topic (`auto.offset.reset = earliest`). The
latter can occur with periodically and unfortunately is a reoccuring theme of
bug (see: KAFKA-9824, KAFKA-9543, KAFKA-9807, KAFKA-9835, etc.)

Addresses #714